naviguide_workspace/requirements.txt
# NAVIGUIDE Workspace — Python dependencies
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic>=2.0.0
python-dotenv==1.0.0

# Maritime routing
searoute==1.4.3
geojson==3.1.0

# LangGraph multi-agent framework
langgraph>=0.2.0
langchain-core>=0.3.0

# LLM provider — Nova 2 Lite + Claude fallback (Bedrock)
<<<<<<< HEAD
google-generativeai>=0.8

=======
boto3>=1.35.0
langchain-aws>=0.2.0
anthropic>=0.40.0
>>>>>>> f4d5b955cff48fe4171a3ea152d75be0dbcc5213

# Data processing
numpy>=1.24.0

# Polar API — PDF parsing & file upload
pdfplumber>=0.11.0
python-multipart>=0.0.9
openpyxl>=3.1.0
pandas>=2.0.0

# Polar API — OCR fallback for raster-scan PDFs (requires: brew install tesseract poppler)
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0
